use oxforddictionaries API
add synonym retrieval

+++ chitchat.py
@@ -1,9 +1,9 @@
 import asyncio
 
 
-async def greet_back(message, text="Hello!"):
+async def send_message(message, text="Hello!", delete_in=False):
     print("received a message", '-', message.content, '\n', message)
     msg = await message.channel.send(text)
-    await asyncio.sleep(3)
-    await msg.delete()
-    print("msg deleted!")
+    if delete_in:
+        await asyncio.sleep(delete_in)
+        await msg.delete()
